deep-learning/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.7.2
reportlab==4.0.4
requests==2.31.0